pyproject.toml
[project]
name = "qmi_etl"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster",
    "dagster-cloud",
    "boto3",
    "pandas",
    "matplotlib",
    "sqlalchemy",
    "pymssql",
    "dagster-dbt",
    "dbt-core>=1.8,<2",
    "dbt-bigquery>=1.8,<2",
    "google-cloud-bigquery",
    "pyarrow",
    "db-dtypes",
]

[dependency-groups]
dev = [
    "dagster-dg-cli>=1.11.11",
    "dagster-webserver",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.scripts]
dbt-build = "dbt: run --project-dir src/qmi_etl/datawarehouse_2 --profiles-dir src/qmi_etl/datawarehouse_2 build"
dbt-parse = "dbt: run --project-dir src/qmi_etl/datawarehouse_2 --profiles-dir src/qmi_etl/datawarehouse_2 parse"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.dagster]
module_name = "qmi_etl.definitions"
code_location_name = "qmietl"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "qmi_etl"
